ConvertedJavaFXPanel: move to gui3 package
gui3 included "conversion support" code.
